VMware Information memory variable changed from int to long
git-svn-id: http://svn.opennebula.org/one/trunk@768 3034c82b-c49b-4eb3-8279-a7acafdc01c0

+++ src/im_mad/vmware/OneImVmware.java
@@ -156,8 +156,8 @@ class OneImVmware extends Thread
                        
                           // Static Information   
                        
-                              int totalMemory = 
-                                     Integer.parseInt(gP.getObjectProperty("hardware.memorySize").toString().trim());
+                              long totalMemory = 
+                                     Long.parseLong(gP.getObjectProperty("hardware.memorySize").toString().trim());
                               totalMemory /= 1024;   
                               
                               response = response + ",TOTALMEMORY=" + totalMemory;
